pyproject.toml
[project]
name = "chatbot-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.15.0",
    "dateparser>=1.2.1",
    "email-validator>=2.0",
    "fastapi>=0.115.12",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.4",
    "langchain-groq>=0.3.2",
    "langchain-qdrant>=0.2.0",
    "langchain-redis>=0.2.1",
    "langchain-text-splitters>=0.3.8",
    "pandas>=2.2.3",
    "pydantic-settings>=2.9.1",
    "pymongo[srv]>=4.13.0",
    "pymupdf>=1.26.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
